Made all file moves for upload checked. This means the file move happens asynchronously then the destination is checked until the size equals what is expected.
Added delay of 500ms to the static method WaitForFileSizeMatch since it was doing a CPU spin using only the Task.Yield() method.

Changed WaitForFileSize match to look for an exact match.

The upload handler now correctly looks for \n after the payload header instead of \r.

The upload handler now uses a more descriptive name for the temporary file in the event things fail. It is now possible to determine what each temporary file was supposed to be in the event it is needed.

The upload handler now does considerable work to ensure each file move is valid. This is the handle network drive setups where even though the OS routine for copy and move will return -- it still needs to be verified to have happened.

The upload handler now calculates the security identifier for each item before writing it to the journal. It was incorrect left blank previously.

Overflow and underflow checking enabled for DEBUG builds, at the least.

Writing to the metajournal now correctly updates the in memory listing. Previously the journal was correctly written but the change was not reflected until the process restarted and re-read the journal.

SempahoreSlim was replaced by AutoResetEvent in DoubleEndedStream. This was a more appropriate usage and it also fixed having to handle an exception. It is the equivalent of a sempahore with a maximum count of 1.

DoubleEndedStream.WaitForRead has an asynchornous counterpart created called WaitForReadAsync.

The async runner task in HTTPClient now has an abort signal called `runner_abort` which is set when a handler throws an exception. This solves the problem of defunt runners and handlers living forever waiting for something that will never happen.